separated into components and added minimal tests for each
